requirements.txt
certifi==2022.9.24
Django==2.2.5
et-xmlfile==1.1.0
interchange==2021.0.4
kafka-python==2.0.2
monotonic==1.6
neo4j==4.4.2
numpy==1.23.3
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
pansi==2020.7.3
peewee==3.14.10
py2neo==2021.2.3
Pygments==2.13.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==5.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
xlrd==1.2.0
